bisect: teach the `bisect--helper` command to show the correct usage strings
In preparation for `bisect--helper` graduating to become the actual
`bisect` command, we hereby teach it to print the very same usage
strings as the scripted `git-bisect.sh` does.

With this patch, the `bisect--helper` command is able to do everything
that the `git-bisect.sh` script could, leaving as last step only to
retire that script at long last, which we will do in the next commit.

Note: Since we cannot use the `parse-options` API to handle the
subcommands of `git bisect` anyway, we no longer use it even just to
show the usage string anymore, either.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>