usage: add warn_once() helper for repeated warnings
The previous change added a warning when valid_remote() detects
credentials in the URL. Since remotes are validated multiple times per
process, this causes multiple warnings to print.

To avoid these kinds of repeated, advisory warnings, create a new
warn_once() helper that behaves the same as warning(), but only after
formatting the output string and adding it to a strset. If that addition
signals that the string already exists in the strset, then do not print
the warning.

In the case of the credentials in a URL, the existing test demonstrates
this per-process limitation: 'git clone' runs 'git-remote-curl' as a
child process, giving two messages. This is an improvement over the
previous six messages.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>